all deps should be optional now
