+++ Cargo.toml
@@ -15,35 +15,48 @@ categories    = ["asynchronous", "network-programming", "parser-implementations"
 rustdoc-args = ["--cfg", "docsrs"]
 all-features = true
 
+[features]
+default = []
+all_docs = []
+async = [
+    "log",
+    "futures-lite",
+    "pin-project-lite",
+    "async-dup",
+    "async-channel",
+    "futures-timer",
+    "fastrand",
+]
+
 [dependencies]
 # logging support
-log = "0.4.11"
+log = { version = "0.4.11", optional = true }
 
 # just the futures traits
-futures-lite = "1.0.0"
+futures-lite = { version = "1.0.0", optional = true }
 
 # field pin projection
-pin-project-lite = "0.1.7"
+pin-project-lite = { version = "0.1.7", optional = true }
 
 # cloneable async writes
-async-dup = "1.2.1"
+async-dup = { version = "1.2.1", optional = true }
 
 # message passing
-async-channel = "1.4.1"
+async-channel = { version = "1.4.2", optional = true }
 
 # for timing out futures
-futures-timer = "3.0.2"
+futures-timer = { version = "3.0.2", optional = true }
 
 # for 'fairness' in the main loop
-fastrand = "1.3.4"
+fastrand = { version = "1.3.5", optional = true }
 
-# serializion/deserializion support
-serde = { version = "1.0.115", optional = true, features = ["derive"] }
+# for optional serialization and deserialization
+serde = { version = "1.0.115", features = ["derive"], optional = true }
 
 # optional runtimes (for TcpStream)
 # these use the futures AsyncWrite+AsyncRead
-async-io  = { version = "0.2.5", optional = true }
-smol      = { version = "0.4.1", optional = true }
+async-io  = { version = "0.2.7", optional = true }
+smol      = { version = "0.4.3", optional = true }
 async-tls = { version = "0.9.0", default-features = false, features = ["client"], optional = true } 
 # TODO look into what their features do. the ones they have enabled by default seem important
 async-std = { version = "1.6.3", optional = true }
@@ -52,7 +65,7 @@ async-std = { version = "1.6.3", optional = true }
 tokio            = { version = "0.2.22", features = ["net"], optional = true } 
 tokio-util       = { version = "0.3.1", features = ["compat"], optional = true }
 
-tokio-rustls     = { version = "0.14.0", optional = true }
+tokio-rustls     = { version = "0.14.1", optional = true }
 webpki-roots     = { version = "0.20.0", optional = true }
 
 tokio-native-tls = { version = "0.1.0", optional = true }
@@ -60,29 +73,31 @@ native-tls       = { version = "0.2.4", optional = true }
 
 [dev-dependencies]
 anyhow         = "1.0.32"
-async-executor = { version = "0.2.0", default-features = false }
-serde_json     = { version = "1.0.57", default-features = false, features = ["alloc"] }
+async-executor = { version = "0.2.1", default-features = false }
+serde_json     = "1.0.57"
 rmp-serde      = "0.14.4"
 
 [[example]]
 name = "message_parse"
+required-features = ["async"]
 
 [[example]]
 name = "smol_demo"
-required-features = ["smol"]
+required-features = ["smol", "async"]
 
 [[example]]
 name = "async_io_demo"
-required-features = ["async-io"]
+required-features = ["async-io", "async"]
 
 [[example]]
 name = "async_std_demo"
-required-features = ["async-std", "async-std/attributes"]
+required-features = ["async-std", "async-std/attributes", "async"]
 
 [[example]]
 name = "tokio_demo"
-required-features = ["tokio/full", "tokio-util"]
+required-features = ["tokio/full", "tokio-util", "async"]
 
 [[example]]
 name = "simple_bot"
-required-features = ["smol"]
+required-features = ["smol", "async"]
+
