+++ src/runner/async_runner.rs
@@ -2,14 +2,14 @@ use crate::{
     channel::Receiver,
     commands,
     connector::Connector,
-    decoder::{AsyncDecoder, DecodeError},
-    encoder::{AsyncEncoder, Encodable},
+    decoder::AsyncDecoder,
+    encoder::AsyncEncoder,
     messages::{Capability, Commands, MessageId},
     rate_limit::{RateClass, RateLimit},
     twitch::UserConfig,
     util::{Notify, NotifyHandle},
     writer::{AsyncWriter, MpscWriter},
-    FromIrcMessage,
+    DecodeError, Encodable, FromIrcMessage, IrcMessage,
 };
 
 use super::{
@@ -163,7 +163,7 @@ impl AsyncRunner {
         log::debug!("joining '{}'", channel);
         self.encoder.encode(commands::join(channel)).await?;
 
-        let channel = crate::commands::Channel(channel).to_string();
+        let channel = crate::commands::Channel::new(channel).to_string();
         log::debug!("waiting for a response");
 
         let mut queue = VecDeque::new();
@@ -211,7 +211,7 @@ impl AsyncRunner {
         log::debug!("leaving '{}'", channel);
         self.encoder.encode(commands::part(channel)).await?;
 
-        let channel = crate::commands::Channel(channel).to_string();
+        let channel = crate::commands::Channel::new(channel).to_string();
         log::debug!("waiting for a response");
 
         let mut queue = VecDeque::new();
@@ -320,8 +320,9 @@ impl AsyncRunner {
             Left(Left(Right(Some(write_data)))) => {
                 // TODO provide a 'bytes' flavored parser
                 let msg = std::str::from_utf8(&*write_data).map_err(Error::InvalidUtf8)?;
-                let msg = crate::irc::IrcMessage::parse(crate::MaybeOwned::Borrowed(msg))
+                let res = crate::irc::parse_one(&msg) //
                     .expect("encoder should produce valid IRC messages");
+                let msg = res.1;
 
                 if let crate::irc::IrcMessage::PRIVMSG = msg.get_command() {
                     if let Some(ch) = msg.nth_arg(0) {
@@ -533,7 +534,7 @@ impl AsyncRunner {
         let mut our_name = None;
 
         let identity = loop {
-            let msg = decoder.read_message().await?;
+            let msg: IrcMessage<'_> = decoder.read_message().await?;
 
             // this should always be infallible. its not marked infallible
             // because of the 'non-exhaustive' attribute
