+++ src/twitch/mod.rs
@@ -16,10 +16,12 @@ pub mod color;
 #[doc(inline)]
 pub use color::Color;
 
+#[allow(dead_code)]
 pub(crate) fn parse_emotes(input: &str) -> Vec<Emotes> {
     Emotes::parse(input).collect()
 }
 
+#[allow(dead_code)]
 pub(crate) fn parse_badges(input: &str) -> Vec<Badge<'_>> {
     input.split(',').filter_map(Badge::parse).collect()
 }
@@ -29,6 +31,7 @@ pub(crate) fn parse_emotes_iter(input: &str) -> impl Iterator<Item = Emotes> + '
     Emotes::parse(input)
 }
 
+#[allow(dead_code)]
 pub(crate) fn parse_badges_iter(input: &str) -> impl Iterator<Item = Badge<'_>> + '_ {
     input.split(',').filter_map(Badge::parse)
 }
